+++ plugins/bigquery/dbt/adapters/bigquery/impl.py
@@ -612,6 +612,10 @@ class BigQueryAdapter(BaseAdapter):
         if dotted_column_name in dbt_columns:
             column_config = dbt_columns[dotted_column_name]
             bq_column_dict['description'] = column_config.get('description')
+            if column_config.get('policy_tags'):
+                bq_column_dict['policyTags'] = {
+                    'names': column_config.get('policy_tags')
+                }
 
         new_fields = []
         for child_col_dict in bq_column_dict.get('fields', list()):
@@ -627,7 +631,7 @@ class BigQueryAdapter(BaseAdapter):
         return bq_column_dict
 
     @available.parse_none
-    def update_column_descriptions(self, relation, columns):
+    def update_columns(self, relation, columns):
         if len(columns) == 0:
             return